+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 `spawn-random` follows [Semantic Versioning][1].
 
+## 2.0.0 - 2018-01-13
+
+### Changed
+
+* `fs.readdir()` errors are now properly passed back to Express
+
+### Fixed
+
+* The `fallthrough` option is no longer overwritten with `true`
+* Providing an empty directory no longer crashes the module
+
 ## 1.0.0 - 2017-06-24
 
 ### Added